zzmp pr feedback
